Add basic authorization

+++ src/movement-logs/movement-logs.controller.ts
@@ -6,14 +6,21 @@ import {
   Patch,
   Param,
   Delete,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { MovementLogsService } from './movement-logs.service';
 import { CreateMovementLogDto } from './dto/create-movement-log.dto';
 import { UpdateMovementLogDto } from './dto/update-movement-log.dto';
+import { AuthorizationService } from 'src/authorization/authorization.service';
+import { CurrentUser } from 'src/authentication/decorators/current-user.decorator';
+import UserSessionInfo from 'src/authentication/entities/UserSessionInfo';
 
 @Controller('movement-logs')
 export class MovementLogsController {
-  constructor(private readonly movementLogsService: MovementLogsService) {}
+  constructor(
+    private readonly movementLogsService: MovementLogsService,
+    private readonly authorizationService: AuthorizationService,
+  ) {}
 
   /**
    * Creates a movement log.
@@ -21,7 +28,19 @@ export class MovementLogsController {
    * Only authorized if logged-in user is author of the movement specified in the request body.
    */
   @Post()
-  create(@Body() createMovementLogDto: CreateMovementLogDto) {
+  async create(
+    @CurrentUser() user: UserSessionInfo,
+    @Body() createMovementLogDto: CreateMovementLogDto,
+  ) {
+    const isAuthorized =
+      await this.authorizationService.userHasAccessToMovement(
+        user.id,
+        createMovementLogDto.movementId,
+      );
+    if (!isAuthorized) {
+      throw new UnauthorizedException();
+    }
+
     return this.movementLogsService.create(createMovementLogDto);
   }
 
@@ -31,7 +50,13 @@ export class MovementLogsController {
    * Only authorized if logged-in user is author of the movement.
    */
   @Get('movement/:id')
-  findAll(@Param('id') id: number) {
+  async findAll(@CurrentUser() user: UserSessionInfo, @Param('id') id: number) {
+    const isAuthorized =
+      await this.authorizationService.userHasAccessToMovement(user.id, id);
+    if (!isAuthorized) {
+      throw new UnauthorizedException();
+    }
+
     return this.movementLogsService.findAll(id);
   }
 
@@ -41,7 +66,13 @@ export class MovementLogsController {
    * Only authorized if logged-in user is author of the movement associated with the movement log.
    */
   @Get(':id')
-  findOne(@Param('id') id: number) {
+  async findOne(@CurrentUser() user: UserSessionInfo, @Param('id') id: number) {
+    const isAuthorized =
+      await this.authorizationService.userHasAccessToMovementLog(user.id, id);
+    if (!isAuthorized) {
+      throw new UnauthorizedException();
+    }
+
     return this.movementLogsService.findOne(id);
   }
 
@@ -51,10 +82,17 @@ export class MovementLogsController {
    * Only authorized if logged-in user is author of the movement associated with the movement log.
    */
   @Patch(':id')
-  update(
+  async update(
+    @CurrentUser() user: UserSessionInfo,
     @Param('id') id: number,
     @Body() updateMovementLogDto: UpdateMovementLogDto,
   ) {
+    const isAuthorized =
+      await this.authorizationService.userHasAccessToMovementLog(user.id, id);
+    if (!isAuthorized) {
+      throw new UnauthorizedException();
+    }
+
     return this.movementLogsService.update(id, updateMovementLogDto);
   }
 
@@ -64,7 +102,13 @@ export class MovementLogsController {
    * Only authorized if logged-in user is author of the movement associated with the movement log.
    */
   @Delete(':id')
-  remove(@Param('id') id: number) {
+  async remove(@CurrentUser() user: UserSessionInfo, @Param('id') id: number) {
+    const isAuthorized =
+      await this.authorizationService.userHasAccessToMovementLog(user.id, id);
+    if (!isAuthorized) {
+      throw new UnauthorizedException();
+    }
+
     return this.movementLogsService.remove(id);
   }
 }
