+++ src/movements/movements.controller.ts
@@ -6,29 +6,39 @@ import {
   Patch,
   Param,
   Delete,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { MovementsService } from './movements.service';
 import { CreateMovementDto } from './dto/create-movement.dto';
 import { UpdateMovmenentDto } from './dto/update-movement.dto';
+import { CurrentUser } from 'src/authentication/decorators/current-user.decorator';
+import UserSessionInfo from 'src/authentication/entities/UserSessionInfo';
+import { AuthorizationService } from 'src/authorization/authorization.service';
 
 @Controller('movements')
 export class MovementsController {
-  constructor(private readonly movementsService: MovementsService) {}
+  constructor(
+    private readonly movementsService: MovementsService,
+    private readonly authorizationService: AuthorizationService,
+  ) {}
 
   /**
    * Creates a movement and sets the currently logged-in user as author.
    */
   @Post()
-  create(@Body() createMovementDto: CreateMovementDto) {
-    return this.movementsService.create(createMovementDto);
+  create(
+    @CurrentUser() user: UserSessionInfo,
+    @Body() createMovementDto: CreateMovementDto,
+  ) {
+    return this.movementsService.create(createMovementDto, user.id);
   }
 
   /**
    * Get all movements for the currently logged-in user.
    */
   @Get()
-  findAll() {
-    return this.movementsService.findAll();
+  findAll(@CurrentUser() user: UserSessionInfo) {
+    return this.movementsService.findAll(user.id);
   }
 
   /**
@@ -37,7 +47,13 @@ export class MovementsController {
    * Only authorized if logged-in user is author of requested movement.
    */
   @Get(':id')
-  findOne(@Param('id') id: number) {
+  async findOne(@CurrentUser() user: UserSessionInfo, @Param('id') id: number) {
+    const isAuthorized =
+      await this.authorizationService.userHasAccessToMovement(user.id, id);
+    if (!isAuthorized) {
+      throw new UnauthorizedException();
+    }
+
     return this.movementsService.findOne(id);
   }
 
@@ -47,10 +63,17 @@ export class MovementsController {
    * Only authorized if logged-in user is author of requested movement.
    */
   @Patch(':id')
-  update(
+  async update(
+    @CurrentUser() user: UserSessionInfo,
     @Param('id') id: number,
     @Body() updateMovementDto: UpdateMovmenentDto,
   ) {
+    const isAuthorized =
+      await this.authorizationService.userHasAccessToMovement(user.id, id);
+    if (!isAuthorized) {
+      throw new UnauthorizedException();
+    }
+
     return this.movementsService.update(id, updateMovementDto);
   }
 
@@ -60,7 +83,13 @@ export class MovementsController {
    * Only authorized if logged-in user is author of requested movement.
    */
   @Delete(':id')
-  remove(@Param('id') id: number) {
+  async remove(@CurrentUser() user: UserSessionInfo, @Param('id') id: number) {
+    const isAuthorized =
+      await this.authorizationService.userHasAccessToMovement(user.id, id);
+    if (!isAuthorized) {
+      throw new UnauthorizedException();
+    }
+
     return this.movementsService.remove(id);
   }
 }
